nl2sql_v3/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nl2sql_v3"
version = "0.1.0"
description = "Natural Language to SQL Table Retrieval CLI Tool"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "elasticsearch>=8.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
nl2sql_v3 = "nl2sql_v3.cli:main"
